Fix doubled/ghost notes in MIDI output
Two bugs:

1. MidiOutputManager::flush used swap-remove to delete events from the
ring buffer. When an event at logical position i was removed, the last
element was copied to position i, but the source slot was not cleared.
When writePos later wrapped and readHead advanced into that dead slot,
the original event was emitted a second time — producing a duplicate
note-on/off pair. Fixed by replacing swap-remove with a linear compaction
pass: events not in the current window are shifted to the front of the
active region, leaving no dead copies. readPos is unchanged; writePos is
maintained implicitly as readPos+count.

2. lastNoteOnSample was stored as int, but noteOnSample is int64. For
sessions longer than ~13 hours the cast overflows; for shorter sessions
with large totalSamplesPlayed values the truncation causes slide notes to
schedule their note-on at a wrong absolute position. Changed the member
type to int64 and removed the narrowing cast.

+++ src/LockEngine.cpp
@@ -172,7 +172,7 @@ void LockEngine::processStep(int step, int64 stepSamplePos, double sampleRate,
     out.scheduleNote(noteOnSample, noteOff,
                      params.outputChannel, midiNote,
                      finalVel, bendVal, bendOffSamples);
-    lastNoteOnSample = (int)noteOnSample;
+    lastNoteOnSample = noteOnSample;
 }
 
 // ─── process ──────────────────────────────────────────────────────────────────
